cdk/requirements.txt
aws-wsgi>=0.2.7
aws-cdk-lib==2.53.0
constructs>=10.0.0,<11.0.0
pandas>=2.2
Flask>=3.0.3
fastavro==0.7.8